v.0.7 - Alpha Seven
* Rebuilt interface so that the main window can be resized
* Rebuilt custom archive window so it can be resized
* Fixed a bug that crashed MASGAU when trying to restore an archive to
steam when no Steam users had been detected
* Added a status bar to replace the splash
* Added file size to reports in analyzer
* Added a UAC indicator to reports in analyzer
* Made analyzer's report window resizable
* Reworked most of the internal cross-thread stuff. More stable, and
allows me to do more with the interface.
* Updated Saved Games folder detection to accommodate custom locations
(just for you cammkelly!)
* Changed all links so they open with the default web browser
* Changed Task tosuppresss dialogs when running in non-interactive mode
* Thanks to a suggestion from VincentV137, MASGAU now has a save
profile to back up a custom.xml file in MASGAU's install folder
* Added a new system.xml for holding definitions of system data
* Added profile for backing up a user's Games for Windows account
information (for Arkham)
* Disabled the purge command for non-windows games
* Little theming tweaks to the interface
* Changed the detected games list to make use of check boxes
* Added support for Windows 7 tasbar progress bars, courtesy of Windows
7 Progress Bars from http://wyday.com/windows-7-progress-bar/
* Blatantly flaunted Microsoft progress bar color guidelines
* Updated monitor so that the detected games list also functions as a
checklist to disable end enable games
* Fixed a bug that caused mods to lose user information
* Inserted a delay into the backup method to allow 7-zip time to finish
working (solves can't write to file error)
* Changed monitoredirectionon to clear out all of it'arrays liststs
properly
* Changed shortcut scanning to deal witarget lessess shortcuts
* Fixed a bug that caused start menu shortcuts to produce incorrect
paths
* Little things here and there that I don't remember

v.0.6 - Sixth Alpha
* Fixed a bug where XP would be detected as Vista
* Fixed a related bug that caused LocalAppData and AppData to
malfunction under XP
* Eliminated the guesswork from user folders, all user folders are now
extracted from each user's hive file for utmost accuracy
* Changed alt paths so on startup if an alt path isn't actually present
it isn't eliminated, just ignored
* Changed restore process so if a folder is being created in a public
location, it will have it's permissions set so everyone can write there
* Added support for new relative path %SAVEDGAMES%, which points to the
user's Saved Games folder under Vista and 7
* Fixed a bug preventing games installed onto the root of a disc from
detecting
* Fixed 2 bugs that prevented monitor from dealing with files that only
appeared momentarily
* Added a small XML file to all backups so that their associated game
and creator can be identified regardless of file name (this means that
you can rename the archives to whatever you want as long as it still
has the .gb7 extension)
* Fixed a bug that prevented monitor from saving files that were at the
moment still in use
* Added failsafe to monitor in case 7-zip is not installed
* Fixed a bug that causes monitor to not shut down all of it's threads
* Added ability to specify file name during custom backup creation
* Fixed a bug that prevented restoring most games
* Changed the path handler to use SubStr instead of Replace
* Added ability to restore many saves without requiring game detection
* Changed Settings tab so the Start monitor on Login check box will be
disabled if the monitor isn't installed
* Changed detected game list to sort by Name first rather than Platform
* Fixed code for games that depend on other games so that it actually
works\
* Added a sub-menu to monitor that shows all the detected games. Handy!
* Fixed a bug that caused analyzer to crash when accessing folders and
files with malformed names

v.0.5-1 Quick fix
* Fixed a bug where Playstation game data objects would have null root
lists

v.0.5 - Fifth Alpha
* Changed 7-zip detection to use a registry key
* Fixed a bug where ps1 saves would only use the ps3 export root
* Fixed a bug where PSP and PS1 saves tried to back up the entire save
folder. Eek.
* Fixed a bug that kept the previous one in check by not letting any
playstation saves back up.
* Fixed a bug that caused playstation saves to appear multiple times in
the Custom backup dialog, could potentially have affected Windows saves
as well
* Fixed a bug that caused the restore dialog to launch twice
* Added a new component for generating save game reports
* Switched to using NSIS instead of MSI for the installer
* Added ability to specify the Windows version a root is for
* Installer now automatically detects if .NET and 7-zip are installed,
and will offer to automatically install them if not found

+++ Masgau/RestoreHandler.cs
@@ -5,18 +5,23 @@ using System.Text;
 using System.IO;
 using System.Diagnostics;
 using System.Windows.Forms;
+using System.Xml;
+using System.Security.AccessControl;
+using wyDay.Controls;
 
 namespace Masgau
 {
     public class RestoreHandler
     {
-        public ArrayList backups;
+        public Dictionary<string, backup_holder> backups = new Dictionary<string,backup_holder>();
+        //public ArrayList backups;
         public string backup_path = null, steam_path = null;
         public PathHandler paths;
         private string temp;
         private bool all_users_mode = false, ready = false;
         private SecurityHandler red_shirt = new SecurityHandler();
         Process unzipper = new Process();
+        invokes invokes = new invokes();
 
         public RestoreHandler(string from_here, string back_this)
         {
@@ -27,7 +32,8 @@ namespace Masgau
                 }
             }
 
-            string path = Environment.GetEnvironmentVariable("PROGRAMFILES") + "\\7-Zip\\7z.exe";
+            RegistryManager zip_path = new RegistryManager("SOFTWARE\\7-Zip");
+            string path = zip_path.getValue("Path") + Path.DirectorySeparatorChar + "7z.exe";
             if (File.Exists(path))
             {
                 unzipper.StartInfo.FileName = path;
@@ -41,38 +47,83 @@ namespace Masgau
         }
 
         public void detectBackups(string back_this) {
-            if(backup_path!=null&&Directory.Exists(backup_path)) {
+            if(ready&&backup_path!=null&&Directory.Exists(backup_path)) {
+                if (Directory.Exists(temp))
+                    Directory.Delete(temp,true);
+                Directory.CreateDirectory(temp);
+                unzipper.StartInfo.UseShellExecute = false;
+                unzipper.StartInfo.RedirectStandardOutput = true;
+                unzipper.StartInfo.RedirectStandardError = true;
+                unzipper.StartInfo.WorkingDirectory = temp;
+                unzipper.StartInfo.CreateNoWindow = true;
+
                 backup_holder add_me;
                 string[] hold_me;
-                backups = new ArrayList();
+                backups = new Dictionary<string,backup_holder>();
                 FileInfo[] read_us = new DirectoryInfo(backup_path).GetFiles("*.gb7");
                 if(read_us.Length>0) {
                     foreach(FileInfo read_me in read_us) {
                         add_me.file_date = read_me.LastWriteTime.ToString();
                         add_me.file_name = read_me.Name;
-                        hold_me = add_me.file_name.Replace(".gb7", "").Split('@');
-                        hold_me = hold_me[0].Split('«');
-                        add_me.game_name = hold_me[0].Trim(); ;
-                        if(hold_me.Length>1) {
-                            add_me.owner = hold_me[1].Trim();
+                        add_me.owner = null;
+                        add_me.game_name = null;
+                        unzipper.StartInfo.Arguments = "x \"" + read_me.FullName + "\" \"masgau.xml\"";
+                        unzipper.Start();
+                        Console.WriteLine(unzipper.StandardOutput.ReadLine());
+                        Console.WriteLine(unzipper.StandardError.ReadLine());
+                        unzipper.WaitForExit();
+                        if(File.Exists(Path.Combine(temp, "masgau.xml"))) {
+                            XmlReaderSettings xml_settings = new XmlReaderSettings();
+                            xml_settings.ConformanceLevel = ConformanceLevel.Document;
+                            xml_settings.IgnoreComments = true;
+                            xml_settings.IgnoreWhitespace = true;
+                            XmlReader load_me = XmlReader.Create(Path.Combine(temp, "masgau.xml"),xml_settings);
+                            while (load_me.Read()) {
+    			                if(load_me.NodeType==XmlNodeType.Element) {
+				                    switch (load_me.Name) {
+                                        case "game":
+                                            while(load_me.MoveToNextAttribute()) {
+                                                switch(load_me.Name) {
+                                                    case "name":
+                                                        add_me.game_name = load_me.Value;
+                                                        break;
+                                                }
+                                            }
+                                            break;
+                                        case "owner":
+                                            while(load_me.MoveToNextAttribute()) {
+                                                switch(load_me.Name) {
+                                                    case "name":
+                                                        add_me.owner = load_me.Value;
+                                                        break;
+                                                }
+                                            }
+                                            break;
+                                    }
+                                }
+                            }
+                            load_me.Close();
+                            File.Delete(Path.Combine(temp, "masgau.xml"));
                         } else {
-                            add_me.owner = null;
+                            hold_me = add_me.file_name.Replace(".gb7", "").Split('@');
+                            hold_me = hold_me[0].Split('«');
+                            add_me.game_name = hold_me[0].Trim(); ;
+                            if(hold_me.Length>1) {
+                                add_me.owner = hold_me[1].Trim();
+                            } else {
+                                add_me.owner = null;
+                            }
                         }
-                        backups.Add(add_me);
+                            
+                        if(add_me.game_name!=null)
+                            backups.Add(add_me.file_name,add_me);
+
                     }
                 } 
             }
         }
 
-        public int findBackup(string find_me) {
-            for(int i = 0;i<backups.Count;i++) {
-                if(((backup_holder)backups[i]).file_name==find_me)
-                    return i;
-            }
-            return -1;
-        }
-
-        public int extractBackup(string restore_me, ProgressBar progress) {
+        public int extractBackup(string restore_me, Windows7ProgressBar progress) {
 
             if(canRead(restore_me)) {
                 if (Directory.Exists(temp))
@@ -95,9 +146,9 @@ namespace Masgau
                     output = unzipper.StandardOutput.ReadLine();
                 }
                 string[] hold_this = last_line.Substring(53).Split(' ');
-                progress.Minimum = 0;
-                progress.Value = 1;
-                progress.Maximum = Int32.Parse(hold_this[0]);
+                invokes.setProgressBarMin(progress,0);
+                invokes.setProgressBarValue(progress,1);
+                invokes.setProgressBarMax(progress,Int32.Parse(hold_this[0]));
                 unzipper.Close();
 
                 unzipper.StartInfo.Arguments = "x \"" + restore_me + "\"";
@@ -106,44 +157,105 @@ namespace Masgau
                 output = unzipper.StandardOutput.ReadLine();
                 while(output != null){
                     if(output.StartsWith("Extracting"))
-                        progress.PerformStep();
+                        invokes.performStep(progress);
                     Console.WriteLine(output);
                     output = unzipper.StandardOutput.ReadLine();
                 }
                 unzipper.Close();
+                if(File.Exists(Path.Combine(temp,"masgau.xml")))
+                    File.Delete(Path.Combine(temp,"masgau.xml"));
                 return progress.Value;
             } else {
-                MessageBox.Show("Can't read " + Path.GetFileName(restore_me) + ".","I'm Illiterate!");
+                MessageBox.Show("Can't read " + Path.GetFileName(restore_me) + ".","I'm Illiterate!",MessageBoxButtons.OK,MessageBoxIcon.Error);
                 return -1;
             }
         }
 
-        public void restoreBackup(string put_here, string owner, ProgressBar progress) {
+
+        public void restoreBackup(string put_here, string owner, Windows7ProgressBar progress, IWin32Window parent, string archive_name) {
             DirectoryInfo from_here = new DirectoryInfo(temp);
             DirectoryInfo to_here = new DirectoryInfo(put_here);
 
+
+			if(!to_here.Exists) {
+				try {
+					to_here.Create();
+					DirectorySecurity everyone = to_here.GetAccessControl();
+					string eveyrones_name = @"Everyone";
+					everyone.AddAccessRule(new FileSystemAccessRule(eveyrones_name,FileSystemRights.FullControl,InheritanceFlags.ContainerInherit|InheritanceFlags.ObjectInherit,PropagationFlags.None,AccessControlType.Allow));
+					to_here.SetAccessControl(everyone);
+				} catch {
+				    if(!red_shirt.amAdmin()) {
+				        if (invokes.showConfirmDialog((Form)parent,"Cheese and Crackers","Unable to create output folder:" + Environment.NewLine + to_here + Environment.NewLine + "Would you like to try again as admin?") != DialogResult.No){
+				            ProcessStartInfo superMode = new ProcessStartInfo();
+				            superMode.UseShellExecute = true;
+				            superMode.WorkingDirectory = Application.ExecutablePath;
+				            superMode.FileName = Path.Combine(Path.GetDirectoryName(Application.ExecutablePath),"MasgauTask.exe");
+				            superMode.Arguments = archive_name;
+				            superMode.Verb = "runas";
+				            try {
+				                Process p = Process.Start(superMode);
+				                p.WaitForExit();
+				            }
+				            catch{
+                                invokes.showMessageBox((Form)parent,"People people people get up get up",superMode.FileName + "can't be found.\nYou probably need to reinstall.",MessageBoxButtons.OK,MessageBoxIcon.Error);
+				            }
+				            return;
+				        }
+				    } else {
+                        invokes.showMessageBox((Form)parent,"And that's how it is","Unable to create output folder:" + Environment.NewLine + to_here,MessageBoxButtons.OK,MessageBoxIcon.Error);
+				    }
+				    return;
+				}
+			}
+
             if(canWrite(to_here.FullName)) {
-                if(!copyFolders(from_here, to_here, progress)){
+                if(!copyFolders(from_here, to_here, progress, parent)){
+					Directory.Delete(temp, true);
                     if(!red_shirt.amAdmin()) {
-                        if (MessageBox.Show("Writing to " + to_here + " is denied.\nWould you like to try running as admin?", "Cheese and Crackers", MessageBoxButtons.YesNo) != DialogResult.No){
-                            red_shirt.elevation(null);
+                        if (invokes.showConfirmDialog((Form)parent, "Cheese and Crackers","Writing to " + to_here + " is denied.\nWould you like to try running as admin?") != DialogResult.No){
+							ProcessStartInfo superMode = new ProcessStartInfo();
+							superMode.UseShellExecute = true;
+							superMode.WorkingDirectory = Application.ExecutablePath;
+						    superMode.FileName = Path.Combine(Path.GetDirectoryName(Application.ExecutablePath),"MasgauTask.exe");
+							superMode.Arguments = archive_name;
+							superMode.Verb = "runas";
+							try {
+								Process p = Process.Start(superMode);
+								p.WaitForExit();
+							}
+							catch{
+                                invokes.showMessageBox((Form)parent,"Everybody everybody stand up stand up",superMode.FileName + "can't be found.\nYou probably need to reinstall.",MessageBoxButtons.OK,MessageBoxIcon.Error);
+							}
                             return;
                         }
                     } else {
-                        MessageBox.Show("Writing to " + to_here + " is denied.", "Cheese and Crackers");
+                        invokes.showMessageBox((Form)parent,"Cheese and Crackers","Writing to " + to_here + " is denied.",MessageBoxButtons.OK,MessageBoxIcon.Error);
                     }
                 }
             } else {
+				Directory.Delete(temp, true);
                 if(!red_shirt.amAdmin()) {
-                    if (MessageBox.Show("Writing to " + to_here + " is denied.\nWould you like to try running as admin?", "Cheese and Crackers", MessageBoxButtons.YesNo) != DialogResult.No){
-                        red_shirt.elevation(null);
+                    if (invokes.showConfirmDialog((Form)parent,"Cheese and Crackers","Writing to " + to_here + " is denied.\nWould you like to try running as admin?") != DialogResult.No){
+						ProcessStartInfo superMode = new ProcessStartInfo();
+						superMode.UseShellExecute = true;
+							superMode.WorkingDirectory = Application.ExecutablePath;
+						superMode.FileName = Path.Combine(Path.GetDirectoryName(Application.ExecutablePath),"MasgauTask.exe");
+						superMode.Arguments = archive_name;
+						superMode.Verb = "runas";
+						try {
+							Process p = Process.Start(superMode);
+							p.WaitForExit();
+						}
+						catch{
+                            invokes.showMessageBox((Form)parent,"Everybody everybody testify",superMode.FileName + " can't be found.\nYou probably need to reinstall.",MessageBoxButtons.OK,MessageBoxIcon.Error);
+						}
                     }
                 } else {
-                    MessageBox.Show("Writing to " + to_here + " is denied.", "Cheese and Crackers");
+                    invokes.showMessageBox((Form)parent,"Crackers and Cheese","Writing to " + to_here + " is denied.",MessageBoxButtons.OK,MessageBoxIcon.Error);
                 }
             }
 
-            Directory.Delete(temp, true);
         }
 
         //private bool createPath(string create_me) {
@@ -162,22 +274,34 @@ namespace Masgau
         //    return true;
         //}
 
-        private bool copyFolders(DirectoryInfo from_here, DirectoryInfo to_here, ProgressBar progress) {
+        private bool copyFolders(DirectoryInfo from_here, DirectoryInfo to_here, Windows7ProgressBar progress, IWin32Window parent) {
             DirectoryInfo now_here;
-            try {
-                if(!to_here.Exists)
-                    Directory.CreateDirectory(to_here.FullName);
-                foreach(FileInfo copy_me in from_here.GetFiles()) {
-                    Console.WriteLine("Copying file: " + copy_me.Name);
-                    progress.PerformStep();
-                    File.Copy(copy_me.FullName,Path.Combine(to_here.FullName,copy_me.Name),true);
-                }
-            } catch {
-                return false;
+            if(!to_here.Exists){
+	            try {
+		            Directory.CreateDirectory(to_here.FullName);
+				} catch {
+                    invokes.showMessageBox((Form)parent,"My genetics are repressed","Error while creating " + to_here.FullName,MessageBoxButtons.OK,MessageBoxIcon.Error);
+					return false;
+				}
+	            try {
+				} catch {
+                    invokes.showMessageBox((Form)parent,"My genetics are repressed","Error while settings permissions on " + to_here.FullName,MessageBoxButtons.OK,MessageBoxIcon.Error);
+					return false;
+				}
+			}
+            foreach(FileInfo copy_me in from_here.GetFiles()) {
+				try {
+					Console.WriteLine("Copying file: " + copy_me.Name);
+                    invokes.performStep(progress);
+					File.Copy(copy_me.FullName,Path.Combine(to_here.FullName,copy_me.Name),true);
+				} catch {
+                    invokes.showMessageBox((Form)parent,"We started this Op'ra","Error while copying " + copy_me.FullName + "\nto " + to_here.FullName,MessageBoxButtons.OK,MessageBoxIcon.Error);
+					return false;
+				}
             }
             foreach(DirectoryInfo check_me in from_here.GetDirectories()) {
                 now_here = new DirectoryInfo(Path.Combine(to_here.FullName,check_me.Name));
-                if(!copyFolders(check_me, now_here,progress))
+                if(!copyFolders(check_me, now_here,progress,parent))
                     return false;
             }
             return true;
@@ -217,5 +341,6 @@ namespace Masgau
             }
             return false;
         }
+
     }
 }
